+++ docker/ros_entrypoint.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-set -e
-
-# Source ROS distro environment
-source "/opt/ros/$ROS_DISTRO/setup.bash"
-
-exec "$@"